safari-600.7-branch/LayoutTests/platform/ios-sim/compositing/overflow/automatically-opt-into-composited-scrolling-expected.txt

Iteration 0: Passed
Iteration 0, layer tree
Iteration 1: Passed
Iteration 1, layer tree
Iteration 2: Passed
Iteration 2, layer tree
Iteration 3: Passed
Iteration 3, layer tree
Iteration 4: Passed
Iteration 4, layer tree
Iteration 5: Passed
Iteration 5, layer tree
Iteration 6: Passed
Iteration 6, layer tree
Iteration 7: Passed
Iteration 7, layer tree
Iteration 8: FAILED
Iteration 8, layer tree
Iteration 9: FAILED
Iteration 9, layer tree
Iteration 10: Passed
Iteration 10, layer tree
Iteration 11: Passed
Iteration 11, layer tree
Iteration 12: FAILED
Iteration 12, layer tree
Iteration 13: FAILED
Iteration 13, layer tree
Iteration 14: FAILED
Iteration 14, layer tree
Iteration 15: FAILED
Iteration 15, layer tree
Iteration 16: FAILED
Iteration 16, layer tree
Iteration 17: FAILED
Iteration 17, layer tree
Iteration 18: FAILED
Iteration 18, layer tree
Iteration 19: FAILED
Iteration 19, layer tree
Iteration 20: Passed
Iteration 20, layer tree
Iteration 21: Passed
Iteration 21, layer tree
Iteration 22: Passed
Iteration 22, layer tree
Iteration 23: Passed
Iteration 23, layer tree
Iteration 24: Passed
Iteration 24, layer tree
Iteration 25: Passed
Iteration 25, layer tree
Iteration 26: Passed
Iteration 26, layer tree
Iteration 27: Passed
Iteration 27, layer tree
Iteration 28: Passed
Iteration 28, layer tree
Iteration 29: Passed
Iteration 29, layer tree
Iteration 30: Passed
Iteration 30, layer tree
Iteration 31: Passed
Iteration 31, layer tree
Iteration 32: Passed
Iteration 32, layer tree
Iteration 33: Passed
Iteration 33, layer tree
Iteration 34: FAILED
Iteration 34, layer tree
Iteration 35: FAILED
Iteration 35, layer tree
Iteration 36: Passed
Iteration 36, layer tree
Iteration 37: Passed
Iteration 37, layer tree
Iteration 38: FAILED
Iteration 38, layer tree
Iteration 39: FAILED
Iteration 39, layer tree
Iteration 40: Passed
Iteration 40, layer tree
Iteration 41: Passed
Iteration 41, layer tree
Iteration 42: Passed
Iteration 42, layer tree
Iteration 43: Passed
Iteration 43, layer tree
Iteration 44: Passed
Iteration 44, layer tree
Iteration 45: Passed
Iteration 45, layer tree
Iteration 46: Passed
Iteration 46, layer tree
Iteration 47: Passed
Iteration 47, layer tree
Iteration 48: FAILED
Iteration 48, layer tree
Iteration 49: FAILED
Iteration 49, layer tree
Iteration 50: Passed
Iteration 50, layer tree
Iteration 51: Passed
Iteration 51, layer tree
Iteration 52: Passed
Iteration 52, layer tree
Iteration 53: Passed
Iteration 53, layer tree
Iteration 54: FAILED
Iteration 54, layer tree
Iteration 55: FAILED
Iteration 55, layer tree
Iteration 56: FAILED
Iteration 56, layer tree
Iteration 57: FAILED
Iteration 57, layer tree
Iteration 58: FAILED
Iteration 58, layer tree
Iteration 59: FAILED
Iteration 59, layer tree
Iteration 60: Passed
Iteration 60, layer tree
Iteration 61: FAILED
Iteration 61, layer tree
Iteration 62: Passed
Iteration 62, layer tree
